+++ app/cli/cmd/integration_add_deptrack.go
@@ -29,9 +29,10 @@ func newIntegrationAddDepTrackCmd() *cobra.Command {
 	var allowAutoCreate bool
 
 	cmd := &cobra.Command{
-		Use:     "dependency-track",
-		Aliases: []string{"deptrack"},
-		Short:   "Add Dependency-Track integration ",
+		Use:        "dependency-track",
+		Aliases:    []string{"deptrack"},
+		Short:      "Add Dependency-Track integration ",
+		Deprecated: "use `chainloop integration add` instead",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			fmt.Print("Enter API Token: \n")
 			apiKey, err := term.ReadPassword(syscall.Stdin)
